// src/app/models/IIDU1200ensayos.ts
export interface IIDU1200ensayos {
  id: number;
  referencia: string;
  area: string;
  ensayoParte: number;
  dcf: string;
  serie: string;
  fecha: string;
  aprobado: number;
  observaciones: string;
  falla: number;
  paso: number;
  corrientes: string;
  velocidades: string;
  moduloWifi: number;
  declaradoPorDia: number;
  aprobadoPorDia: number;
}